requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
flash_attn==2.5.8
torch==2.2.0
accelerate==0.31.0
transformers==4.37.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
safetensors>=0.5.2
# Remove potential conflicts
# torchaudio and torchvision are not needed for this implementation